+++ src/Redmine/Api/User.php
@@ -26,11 +26,29 @@ class User extends AbstractApi
      *
      * @return array list of users found
      */
-    public function all(array $params = [])
+    final public function list(array $params = []): array
     {
         $this->users = $this->retrieveData('/users.json', $params);
 
         return $this->users;
+    }
+
+    /**
+     * List users.
+     *
+     * @deprecated since v2.4.0, use list() instead.
+     *
+     * @see http://www.redmine.org/projects/redmine/wiki/Rest_Users#GET
+     *
+     * @param array $params to allow offset/limit (and more) to be passed
+     *
+     * @return array list of users found
+     */
+    public function all(array $params = [])
+    {
+        @trigger_error('`'.__METHOD__.'()` is deprecated since v2.4.0, use `'.__CLASS__.'::list()` instead.', E_USER_DEPRECATED);
+
+        return $this->list($params);
     }
 
     /**